Unify tinyms model definition into model module
